[refactor] グラフ定義を別ファイルに移動
