+++ langgraph.json
@@ -2,6 +2,7 @@
   "graphs": {
     "agent": "./main.py:app"
   },
+  "python_version": "3.12",
   "dependencies": ["./"],
   "env": "./.env"
 }
