+++ pyproject.toml
@@ -10,4 +10,5 @@ dependencies = [
     "langgraph-cli[inmem]>=0.4.0",
     "langgraph-sdk>=0.2.4",
     "pydantic>=2.11.7",
+    "python-dotenv>=1.1.1",
 ]
